filters don't like unsigned values as input :/

+++ modules/cvra_servo/cvra_servo.c
@@ -7,14 +7,14 @@ void cvra_servo_set(void *base, int channel, uint32_t value) {
     IOWR(base, channel, value);
 }
 
-void cvra_servo_set0(void *base, uint32_t value){
-    cvra_servo_set(base, 0, value);
+void cvra_servo_set0(void *base, int32_t value){
+    cvra_servo_set(base, 0, (uint32_t)value);
 }
 
-void cvra_servo_set1(void *base, uint32_t value){
-    cvra_servo_set(base, 1, value);
+void cvra_servo_set1(void *base, int32_t value){
+    cvra_servo_set(base, 1, (uint32_t)value);
 }
 
-void cvra_servo_set2(void *base, uint32_t value){
-    cvra_servo_set(base, 2, value);
+void cvra_servo_set2(void *base, int32_t value){
+    cvra_servo_set(base, 2, (uint32_t)value);
 }
